Creating generalEffectDf() function and support
Also required functionalising getRisk(), previously a per-model-type set of if statements in cIndex(), which itself was generalised by functionalising modelType(), rather than passing model.type to every function (which is of course error-prone anyway). model.type is still not phased out in some functions where eg a data frame is passed and it's not obvious what type of model it came from.