survivalFitBoot for parallel bootstrapping of survival fits
This is included for posterity/future inclusion at this point...it still doesn't work, because (just as in the boot function) the calibration score throws an error and causes that process to crash. I'm about to rewrite it for single-core execution, but just in case this parallel version needs restoring, this is what it looks like.